Each configuration file using its own Google Storage credentials file
Pushing a dummy `credentials.test` with `.json` and `.env` extensions to
the `secrets` folder, so that Travis CI is able to run the tests on its
own, without the need to actually know any of the real secrets used in
runtime.

This commit changes how we get the Google credentials with the Goth
library, we now fetch them from a `GOOGLE_STORAGE_CREDENTIALS`
environment variable that contains the Base64 encoded credentials.

The `GOOGLE_STORAGE_CREDENTIALS` was set on the `env/test.env`
environment file with a dummy Base64 value (but respecting the JSON
structure the Goth authentication library expects), since the
environment variables contained in this file are passed to the container
when executing the `docker-compose.test.yml` docker compose file.