We don't use the "prod.secret.exs" ElixirDrip secret configuration
The `elixir_drip_web` used the `prod.secret.exs` configuration file to
store the secret key base. Since we don't want to use this file, we are
fetching this value from a `SECRET_KEY_BASE` environment variable (with
`System.get_env/2` in every environment except `prod`, in which we run
the release, and so we use the `${SECRET_KEY_BASE}` format).