Passing the Google Storage Credentials as a secret via env vars
We've encoded the Google Storage JSON credentials as Base64 so they can
be easily stored as an environment variable. The library we're using to
handle the Google authentication for us is relying on the existence of a
`GOOGLE_STORAGE_CREDENTIALS` environment var with those credentials.

Travis CI also fills this environment var, but with dummy values (just
the correct expected JSON structure), because Travis CI doesn't need the
actual values.