Custom vm.args used by the release
We're passing the POD_IP and the ERLANG_COOKIE as environment variables
to the `elixir-drip-prod` container that runs the application (the first
comes from the kubernetes info, the latter from k8s secrets).

We changed the release configuration to set the node name and cookie
based on the previous environment variables, through the interpolation
of ${POD_IP} in the node name and ${ERLANG_COOKIE} in the cookie.