Adding labels to the load balancer that creates the endpoints for us
I'm able to manually connect the nodes between themselves. After reading
the `libcluster` documentation again, I've seen that `libcluster` uses
the label selector to query the existing endpoints.

This gave me a clue that we didn't have any labels on the loadbalancer
that's creating the endpoints for us.

After adding the labels, everything worked (because `libcluster` tries
to connect every 5 secs).