Add `MyPlugPipeline`
This module uses the `Plug.Builder` module to create a pipeline of two
Plugs, one being a module Plug (defined elsewhere) and the other being a
function Plug defined in this module. This will not be used in the
ElixirDrip application, and is just here as part of an example present
in the Mastering Elixir book.