Add `login_user_with_pw` function to the Accounts context
This function allows us to provide a username + password pair and check
if the credentials are correct. If they are, this function returns
`{:ok, user}` (with `user` being a %User{} struct). If they aren't, this
function returns `:error`. It's important to note that we're using the
`dummy_checkpw()` function from the `comeonin` library to simulate the
check of a password when the user isn't found, to protect our
application from timing attacks.