Add `get_file_info` function to Storage context
This function is responsible for returning the information corresponding
to a specific file. To implement this, a `get_media` function was
created to get the media based on both the `user_id` and the `media_id`.
There was already a function with this name, which was renamed to
`find_media`.