Add Notifications for File uploads, downloads and shares
This commit adds the ability to render a notification on the application
whenever a file is either uploaded, downloaded or shared. To enable
this, we had to update the final step of the pipelines to call the newly
crated `Notifications` module, which will create an event on the user's
channel.