ElixirDrip.Instrumenter for custom counters and histograms
Introduces the `defmeasured/2` macro to define functions. It then
measures the time the function takes to execute and sends the
corresponding histogram metrics.

ElixirDrip.Instrumenter.setup/0 needs to be called when the application
starts, to register the custom metrics that we'll send to Prometheus.

ElixirDrip.Instrumenter also has some functions to create the counter
and histogram metric names (NB histogram names need to end with the time
units, or else they need to be passed in the metric spec with
the `duration_units` parameter).